refactor(db): extract 7 leaf entity modules — decomposition Wave 1
Second wave of the db.js decomposition. Extracts the low-coupling leaf entities into
server/db/*.js; the barrel re-exports each via `export *`, so no consumer's import
changes. Behavior-preserving (bodies moved verbatim).

Modules (34 functions): spend(3), config(6), bugs(6), feedback(5), events(3),
widgets(4), skills(7). Each imports only ./core.js — acyclic (a db/*.js never imports
the ../db.js barrel).

Subtlety handled: `export *` re-exports moved names to CONSUMERS but does not bind them
in db.js's own lexical scope. Five resident db.js functions still called moved helpers
by bare name (listEvents, listBugs, countBugs, getSleepMode, listInstanceConfig) — now
imported explicitly into db.js from their new modules. (All 34 moved names were grepped
for stale callers, not just the tested ones — an untested call site would be a latent
regression the manifest gate structurally cannot catch.)

db.js: 4257 -> 3995 (-262). getWidget left in db.js for now (not in the blueprint's
widgets.js) — flagged for a later wave.

Gate: db-manifest --check green (308 unchanged); full suite 45 files / 310 tests.

Executor: GLM-5.2 (agentic, gate-driven). Reviewed + verified by m5Max.
